Collect metrics on memory usage
Fix issue with less than or equal to 8byte allocations - use singly linked list for freelist bucket for 8bytes allocs[it will be bumped up to 16bytes]. This is because we need 8 bytes overhead to store the size when sending to user. We reuse the same 8bytes to store pointer to next 8byte chunk in the freelist bucket. We don"t need to store size while the chunk is in the freelist since this can be calculated